Add InfiniDB & MariaDB ColumnStore
Added what I know of the InfiniDB and ColumnStore forks.

InfiniDB was a fork of MySQL to add a columnar storage engine, this
required a lot of modifications to MySQL's internals, hence the fork.

MariaDB ColumnStore was a fork of InfiniDB combined with a fork of
MariaDB 10.X.

Things get a little complicated around MariaDB ColumnStore 1.4. It is
no longer a fork due to new APIs added in MariaDB. It is now a
separate tree for the engine parts, it has only been released with
MariaDB Enterprise 10.4 due to timing. It is expected that ColumnStore
engine 1.5 will be part of MariaDB 10.5.